Move sleep operation into subdir
