+++ packages/core/src/index.ts
@@ -5,10 +5,11 @@ import { Effection } from './effection';
 export { State, StateTransition } from './state-machine';
 export { createTask, Task, TaskOptions, Controls, getControls } from './task';
 export { Operation, Resource } from './operation';
-export { sleep } from './sleep';
 export { Effection } from './effection';
 export { deprecated } from './deprecated';
 export { Deferred } from './deferred';
+
+export { sleep } from './operations/sleep';
 
 export function run<TOut>(operation?: Operation<TOut>, options?: TaskOptions): Task<TOut> {
   return Effection.root.spawn(operation, options);
